Consolidate duplicate console report logic to an abstract class
