+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/reporting/console/AbstractFindingsReport.kt
@@ -3,14 +3,12 @@ package io.gitlab.arturbosch.detekt.core.reporting.console
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.ConsoleReport
 import io.gitlab.arturbosch.detekt.api.Detektion
+import io.gitlab.arturbosch.detekt.api.Finding
+import io.gitlab.arturbosch.detekt.api.RuleSetId
 import io.gitlab.arturbosch.detekt.api.SingleAssign
 import io.gitlab.arturbosch.detekt.core.reporting.filterEmptyIssues
 
-/**
- * Contains a clear read of the console report, where each line contains location, messages and issue id.
- * See: https://detekt.github.io/detekt/configurations.html#console-reports
- */
-class ReadableFindingsReport : ConsoleReport() {
+abstract class AbstractFindingsReport : ConsoleReport() {
 
     private var config: Config by SingleAssign()
 
@@ -21,15 +19,12 @@ class ReadableFindingsReport : ConsoleReport() {
     }
 
     override fun render(detektion: Detektion): String? {
-        val issues = detektion.filterEmptyIssues(config)
-        if (issues.isEmpty()) {
+        val findings = detektion.filterEmptyIssues(config)
+        if (findings.isEmpty()) {
             return null
         }
-        return buildString {
-            issues.values.flatten().forEach { finding ->
-                append("${finding.location.compact()}: ${finding.messageOrDescription()} [${finding.issue.id}]")
-                appendLine()
-            }
-        }
+        return render(findings)
     }
+
+    abstract fun render(findings: Map<RuleSetId, List<Finding>>): String
 }
