+++ detekt-core/src/main/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.ConsoleReport
@@ -1,6 +1,6 @@
 io.gitlab.arturbosch.detekt.core.reporting.console.ComplexityReport
 io.gitlab.arturbosch.detekt.core.reporting.console.FindingsReport
 io.gitlab.arturbosch.detekt.core.reporting.console.FileBasedFindingsReport
-io.gitlab.arturbosch.detekt.core.reporting.console.ReadableFindingsReport
+io.gitlab.arturbosch.detekt.core.reporting.console.LiteFindingsReport
 io.gitlab.arturbosch.detekt.core.reporting.console.NotificationReport
 io.gitlab.arturbosch.detekt.core.reporting.console.ProjectStatisticsReport
